+++ common/idstring.cc
@@ -1,0 +1,51 @@
+/*
+ *  nextpnr -- Next Generation Place and Route
+ *
+ *  Copyright (C) 2018  Clifford Wolf <clifford@symbioticeda.com>
+ *  Copyright (C) 2018  Serge Bazanski <q3k@symbioticeda.com>
+ *
+ *  Permission to use, copy, modify, and/or distribute this software for any
+ *  purpose with or without fee is hereby granted, provided that the above
+ *  copyright notice and this permission notice appear in all copies.
+ *
+ *  THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
+ *  WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF
+ *  MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR
+ *  ANY SPECIAL, DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
+ *  WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN
+ *  ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
+ *  OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
+ *
+ */
+
+#include "idstring.h"
+
+#include "basectx.h"
+
+NEXTPNR_NAMESPACE_BEGIN
+
+void IdString::set(const BaseCtx *ctx, const std::string &s)
+{
+    auto it = ctx->idstring_str_to_idx->find(s);
+    if (it == ctx->idstring_str_to_idx->end()) {
+        index = ctx->idstring_idx_to_str->size();
+        auto insert_rc = ctx->idstring_str_to_idx->insert({s, index});
+        ctx->idstring_idx_to_str->push_back(&insert_rc.first->first);
+    } else {
+        index = it->second;
+    }
+}
+
+const std::string &IdString::str(const BaseCtx *ctx) const { return *ctx->idstring_idx_to_str->at(index); }
+
+const char *IdString::c_str(const BaseCtx *ctx) const { return str(ctx).c_str(); }
+
+void IdString::initialize_add(const BaseCtx *ctx, const char *s, int idx)
+{
+    NPNR_ASSERT(ctx->idstring_str_to_idx->count(s) == 0);
+    NPNR_ASSERT(int(ctx->idstring_idx_to_str->size()) == idx);
+    auto insert_rc = ctx->idstring_str_to_idx->insert({s, idx});
+    ctx->idstring_idx_to_str->push_back(&insert_rc.first->first);
+}
+
+NEXTPNR_NAMESPACE_END
