+++ common/idstring.h
@@ -1,0 +1,73 @@
+/*
+ *  nextpnr -- Next Generation Place and Route
+ *
+ *  Copyright (C) 2018  Clifford Wolf <clifford@symbioticeda.com>
+ *  Copyright (C) 2018  Serge Bazanski <q3k@symbioticeda.com>
+ *
+ *  Permission to use, copy, modify, and/or distribute this software for any
+ *  purpose with or without fee is hereby granted, provided that the above
+ *  copyright notice and this permission notice appear in all copies.
+ *
+ *  THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
+ *  WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF
+ *  MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR
+ *  ANY SPECIAL, DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
+ *  WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN
+ *  ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
+ *  OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
+ *
+ */
+
+#ifndef IDSTRING_H
+#define IDSTRING_H
+
+#include <string>
+#include "nextpnr_namespaces.h"
+
+NEXTPNR_NAMESPACE_BEGIN
+
+struct BaseCtx;
+
+struct IdString
+{
+    int index;
+
+    static void initialize_arch(const BaseCtx *ctx);
+
+    static void initialize_add(const BaseCtx *ctx, const char *s, int idx);
+
+    constexpr IdString() : index(0) {}
+    explicit constexpr IdString(int index) : index(index) {}
+
+    void set(const BaseCtx *ctx, const std::string &s);
+
+    IdString(const BaseCtx *ctx, const std::string &s) { set(ctx, s); }
+
+    IdString(const BaseCtx *ctx, const char *s) { set(ctx, s); }
+
+    const std::string &str(const BaseCtx *ctx) const;
+
+    const char *c_str(const BaseCtx *ctx) const;
+
+    bool operator<(const IdString &other) const { return index < other.index; }
+
+    bool operator==(const IdString &other) const { return index == other.index; }
+
+    bool operator!=(const IdString &other) const { return index != other.index; }
+
+    bool empty() const { return index == 0; }
+};
+
+NEXTPNR_NAMESPACE_END
+
+namespace std {
+template <> struct hash<NEXTPNR_NAMESPACE_PREFIX IdString>
+{
+    std::size_t operator()(const NEXTPNR_NAMESPACE_PREFIX IdString &obj) const noexcept
+    {
+        return std::hash<int>()(obj.index);
+    }
+};
+} // namespace std
+
+#endif /* IDSTRING_H */
