+++ common/nextpnr_assertions.h
@@ -1,0 +1,62 @@
+/*
+ *  nextpnr -- Next Generation Place and Route
+ *
+ *  Copyright (C) 2018  Clifford Wolf <clifford@symbioticeda.com>
+ *  Copyright (C) 2018  Serge Bazanski <q3k@symbioticeda.com>
+ *
+ *  Permission to use, copy, modify, and/or distribute this software for any
+ *  purpose with or without fee is hereby granted, provided that the above
+ *  copyright notice and this permission notice appear in all copies.
+ *
+ *  THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
+ *  WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF
+ *  MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR
+ *  ANY SPECIAL, DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
+ *  WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN
+ *  ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
+ *  OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
+ *
+ */
+
+#ifndef NEXTPNR_ASSERTIONS_H
+#define NEXTPNR_ASSERTIONS_H
+
+#include <stdexcept>
+
+#include "nextpnr_namespaces.h"
+
+NEXTPNR_NAMESPACE_BEGIN
+
+class assertion_failure : public std::runtime_error
+{
+  public:
+    assertion_failure(std::string msg, std::string expr_str, std::string filename, int line);
+
+    std::string msg;
+    std::string expr_str;
+    std::string filename;
+    int line;
+};
+
+NPNR_NORETURN
+inline void assert_fail_impl(const char *message, const char *expr_str, const char *filename, int line)
+{
+    throw assertion_failure(message, expr_str, filename, line);
+}
+
+NPNR_NORETURN
+inline void assert_fail_impl_str(std::string message, const char *expr_str, const char *filename, int line)
+{
+    throw assertion_failure(message, expr_str, filename, line);
+}
+
+#define NPNR_ASSERT(cond) (!(cond) ? assert_fail_impl(#cond, #cond, __FILE__, __LINE__) : (void)true)
+#define NPNR_ASSERT_MSG(cond, msg) (!(cond) ? assert_fail_impl(msg, #cond, __FILE__, __LINE__) : (void)true)
+#define NPNR_ASSERT_FALSE(msg) (assert_fail_impl(msg, "false", __FILE__, __LINE__))
+#define NPNR_ASSERT_FALSE_STR(msg) (assert_fail_impl_str(msg, "false", __FILE__, __LINE__))
+
+#define NPNR_STRINGIFY(x) #x
+
+NEXTPNR_NAMESPACE_END
+
+#endif /* NEXTPNR_ASSERTIONS_H */
