+++ common/relptr.h
@@ -1,4 +1,31 @@
-// This is intended to be included inside arch.h only.
+/*
+ *  nextpnr -- Next Generation Place and Route
+ *
+ *  Copyright (C) 2018  Clifford Wolf <clifford@symbioticeda.com>
+ *
+ *  Permission to use, copy, modify, and/or distribute this software for any
+ *  purpose with or without fee is hereby granted, provided that the above
+ *  copyright notice and this permission notice appear in all copies.
+ *
+ *  THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
+ *  WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF
+ *  MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR
+ *  ANY SPECIAL, DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
+ *  WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN
+ *  ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
+ *  OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
+ *
+ */
+
+#ifndef RELPTR_H
+#define RELPTR_H
+
+#include <cstdint>
+
+#include "nextpnr_assertions.h"
+#include "nextpnr_namespaces.h"
+
+NEXTPNR_NAMESPACE_BEGIN
 
 template <typename T> struct RelPtr
 {
@@ -6,7 +33,7 @@ template <typename T> struct RelPtr
 
     const T *get() const { return reinterpret_cast<const T *>(reinterpret_cast<const char *>(this) + offset); }
 
-    const T &operator[](size_t index) const { return get()[index]; }
+    const T &operator[](std::size_t index) const { return get()[index]; }
 
     const T &operator*() const { return *(get()); }
 
@@ -22,7 +49,7 @@ NPNR_PACKED_STRUCT(template <typename T> struct RelSlice {
 
     const T *get() const { return reinterpret_cast<const T *>(reinterpret_cast<const char *>(this) + offset); }
 
-    const T &operator[](size_t index) const
+    const T &operator[](std::size_t index) const
     {
         NPNR_ASSERT(index < length);
         return get()[index];
@@ -41,3 +68,7 @@ NPNR_PACKED_STRUCT(template <typename T> struct RelSlice {
     RelSlice(const RelSlice &) = delete;
     RelSlice &operator=(const RelSlice &) = delete;
 });
+
+NEXTPNR_NAMESPACE_END
+
+#endif /* RELPTR_H */
